ansible/roles/deploy_app/files/requirements.txt
aiohttp
kafka-python
pymongo